pyproject.toml
[project]
name = "chat-with-jfk-files"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "voynow", email = "voynow99@gmail.com" }]
dependencies = [
    "openai>=1.60.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.7",
    "uvicorn>=0.30.1",
    "pydantic>=2.1.1",
    "gunicorn>=23.0.0",
    "pinecone>=5.4.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false
